Add tutorial component styles
