+++ src/layouts/BaseLayout.astro
@@ -44,7 +44,7 @@ const currentPage = Astro.url.pathname;
 
 <style>
   :root {
-    --left-sidebar-width: 300px;
+    --left-sidebar-width: 400px;
     --right-sidebar-width: 240px;
   }
